% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRTS_score.R
\name{Add_metadata}
\alias{Add_metadata}
\title{Add meta data to the \code{miRTS_score} output}
\usage{
Add_metadata(miR_TS.output, meta_data)
}
\arguments{
\item{miR_TS.output}{the output from miRTS_score()}

\item{meta_data}{the meta data for the \code{Input_df} to miRTS_score();
should include:
Col_names = Sample names used in the \code{Input_df} (e.g., 'S001', 'S002')
sample_type = grouping of the sample (e.g.,'control', 'disease')}
}
\value{
Matrix: samples (rows) ×  tissues+sample_type (columns).
}
\description{
Add meta data to the \code{miRTS_score} output
}
\examples{
dim(miR_TS.output.bkup$proportions); dim(hepatitis_C.meta)
hepatitis_C.output <- miRTS::Add_metadata(miR_TS.output.bkup$proportions, hepatitis_C.meta)
dim(hepatitis_C.output)
}
