% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORT.R
\name{CIBERSORT}
\alias{CIBERSORT}
\title{The CIBERSORT.R is licensed but free of charge for non-commercial use only.
Following registration, the \strong{CIBERSORT.R} file is available via the \href{https://cibersortx.stanford.edu/}{cibersort website}.}
\usage{
CIBERSORT(...)
}
\description{
The CIBERSORT.R is licensed but free of charge for non-commercial use only.
Following registration, the \strong{CIBERSORT.R} file is available via the \href{https://cibersortx.stanford.edu/}{cibersort website}.
}
