% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORT.R
\name{CIBERSORT_download}
\alias{CIBERSORT_download}
\title{A small function to download the CIBERSORT function}
\usage{
CIBERSORT_download()
}
\value{
the cibersortx website.
}
\description{
The CIBERSORT.R is licensed but free of charge for non-commercial use only.
Following registration, the \strong{CIBERSORT.R} file is available via the \href{https://cibersortx.stanford.edu/}{cibersort website}.
}
\examples{
CIBERSORT_download()
}
