% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miR_TS.output.bkup}
\alias{miR_TS.output.bkup}
\title{Output object saved for testing the miRTS_score function}
\format{
An object of class \code{list} of length 3.
}
\usage{
miR_TS.output.bkup
}
\description{
The miR_TS.output.bkup object was archived in this package to
enable the testing of the \code{miRTS_score} function. The
output of the miRTS_score(Input_df = example_counts) should be this object.
}
\details{
This list object includes the following:
\enumerate{
\item the "proportions" is the estimated miR_TS score;
\item the "mix" is the z-scaled input mixture data (\code{Input_df });
\item the "signature" is the z-scaled signature matrix (\code{miRTS_signature_v1})
}
}
\examples{
data(miR_TS.output.bkup)
dim(miR_TS.output.bkup)
}
\keyword{datasets}
