% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRTS_score.R
\name{miRTS_score}
\alias{miRTS_score}
\title{miRNA-based Tissue Signal (miR-TS) scores}
\usage{
miRTS_score(
  Input_df,
  signature_matrix = miRTS::miRTS_signature_v1,
  method = c("cibersort", "xCell2", "MCP-counter"),
  Dtct_cutoff = 0.1,
  Norm = c("TMM", "RLE", "none"),
  ts_miR.df = ts_miRNA,
  outlier_sample.cutoff = NA,
  useAbsolute = TRUE,
  Use_blood = TRUE,
  ...
)
}
\arguments{
\item{Input_df}{Input miRNAs data with miRNA names on the row
and sample names on the column.
Read counts (or similar on an exponential scale) should be used.}

\item{signature_matrix}{matrix: miRNAs x tissues,
defaults to the signature matrix optimized from miRNATissueAtlas2 (PMID: 34570238)}

\item{method}{character: deconvolution method ("cibersort" default)}

\item{Dtct_cutoff}{Detect rate cutoff to include a subset of
miRNAs to be considered as reliably measured, default is 10\%;
note that a miRNA with read count being 0 still inform the miR-TS,
so it is advised to keep this number low (10\% or 20\%) to include
as many signature miRNAs as possible.}

\item{Norm}{Method for normalizing \code{Input_df}, default to TMM.}

\item{ts_miR.df}{a parameter required when using the MCP-counter method;
Default to ts_miRNA, which records the enriched tissues for each miRNAs in the \code{miRTS_signature_v1}}

\item{outlier_sample.cutoff}{Cutoff of the MAD folds of the library sizes
to define outlier samples. Default to NA.}

\item{useAbsolute}{Use Absolute mode or not. Default is TRUE.}

\item{Use_blood}{Use plasma or serum samples or not. Default is TRUE.}

\item{...}{additional arguments passed to the deconvolution engine}
}
\value{
A list, including: proportions (the estimated tissue scores),
mix (the \code{Input_df}), signatures (the \code{signature_matrix})
}
\description{
Compute miRNA-based Tissue Signal (miR-TS) scores from extracellular miRNAs in circulation
}
\examples{
miR_TS.output <- miRTS_score(Input_df = example_counts)
all.equal(miR_TS.output, miR_TS.output.bkup) # miR_TS.output.bkup is the ultimate output saved in the package for testing.
dim(miR_TS.output$proportions)
}
