% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tissues.blood_detect_17}
\alias{Tissues.blood_detect_17}
\title{17 tissue types with miRTS signals detectable in plasma/serum}
\format{
An object of class \code{character} of length 17.
}
\usage{
Tissues.blood_detect_17
}
\description{
17 tissue types with miRTS signals detectable in plasma/serum
}
\details{
A vector of characters. The 17 tissue types with miRTS signals detectable in plasma/serum.
}
\keyword{datasets}
