% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xCell2Object.miRTS_signature}
\alias{xCell2Object.miRTS_signature}
\title{The "xCell2Object" for \code{miRTS_score()} when applying xCell2}
\format{
An object of class \code{xCell2Object} of length 1.
}
\usage{
xCell2Object.miRTS_signature
}
\description{
This xCell2Object was created by applying the xCell2::xCell2Train() function
to the miRTS_signature_v1 with default settings.
}
\details{
When \code{method = "xCell2"} is selected in \code{miRTS_score()}, this xCell2Object is needed as input.
}
\keyword{datasets}
